requirements.txt
openai
langchain
llama-index
python-dotenv
langchain-openai